Organize Braintrust and TRL as integrations
