+++ docs/documentation_home.mdx
@@ -13,7 +13,7 @@ Welcome to the Reward Kit documentation. This guide will help you create, test, 
 - [Evaluation Workflows](developer_guide/evaluation_workflows.mdx): Learn the complete lifecycle from development to deployment.
 - [Dataset Configuration Guide](dataset_configuration_guide.md): Understand how to configure datasets using YAML.
 - [Hydra Configuration for Examples](developer_guide/hydra_configuration.mdx): Learn how Hydra is used for configuration in examples.
-- [Integrating with Braintrust](developer_guide/braintrust_integration.mdx): Bridge Reward Kit with the Braintrust SDK.
+- [Integrating with Braintrust](integrations/braintrust_integration.mdx): Bridge Reward Kit with the Braintrust SDK.
 
 ### Examples
 - [Examples Overview](examples/examples_overview.mdx): Browse available examples and learn how to run them. The primary documentation for each example is its own `README.md` in the `examples/` directory.
