+++ docs/examples/examples_overview.mdx
@@ -36,7 +36,7 @@ Many examples use [Hydra for configuration](./../developer_guide/hydra_configura
 
 *   **TRL Integration Example**:
     *   Shows how to integrate reward-kit functions with the TRL library.
-    *   [View Documentation](./trl_integration/trl_integration_overview.mdx)
+    *   [View Documentation](../integrations/trl_integration_overview.mdx)
 
 *Note: The `examples/metrics/` and `examples/test_tasks/` directories in the root `examples/` folder contain supporting resources and are not standalone documented examples here.*
