+++ docs/integrations/braintrust_integration.mdx
@@ -16,7 +16,7 @@ Convert a Braintrust-style scorer to a Reward Kit reward function using `scorer_
 
 ```python
 from braintrust import Eval
-from reward_kit.adapters.braintrust import scorer_to_reward_fn
+from reward_kit.integrations.braintrust import scorer_to_reward_fn
 
 
 def equality_scorer(input: str, output: str, expected: str) -> float:
