+++ examples/trl_integration/grpo_example.py
@@ -18,7 +18,7 @@ import torch
 # Ensure reward-kit is in the path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 
-from reward_kit.adapters.trl import (  # Import the new generic TRL adapter
+from reward_kit.integrations.trl import (  # Import the new generic TRL adapter
     create_trl_adapter,
 )
 from reward_kit.models import EvaluateResult, MetricResult
