+++ examples/trl_integration/working_grpo_example.py
@@ -22,7 +22,7 @@ logger = logging.getLogger(__name__)
 # Ensure reward-kit is in the path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 
-from reward_kit.adapters.trl import (  # Import the new generic TRL adapter
+from reward_kit.integrations.trl import (  # Import the new generic TRL adapter
     create_trl_adapter,
 )
 from reward_kit.models import (  # RewardOutput, MetricRewardOutput are legacy
