+++ reward_kit/__init__.py
@@ -13,7 +13,7 @@ import warnings
 from .models import EvaluateResult, Message, MetricResult
 from .reward_function import RewardFunction
 from .typed_interface import reward_function
-from .adapters.braintrust import scorer_to_reward_fn, reward_fn_to_scorer
+from .integrations.braintrust import scorer_to_reward_fn, reward_fn_to_scorer
 
 warnings.filterwarnings("default", category=DeprecationWarning, module="reward_kit")
