+++ reward_kit/adapters/braintrust.py
@@ -1,59 +1,9 @@
-"""Adapters for integrating Reward Kit with Braintrust scoring functions."""
+"""Deprecated adapter wrappers for Braintrust.
 
-from typing import Any, Callable, List, Optional
+This module forwards imports to :mod:`reward_kit.integrations.braintrust`.
+"""
 
-from reward_kit.models import EvaluateResult, Message
-from reward_kit.typed_interface import reward_function
+from ..integrations.braintrust import scorer_to_reward_fn, reward_fn_to_scorer
 
-# Type alias for Braintrust scoring functions
-BraintrustScorer = Callable[[Any, Any, Any], float]
-
-
-def scorer_to_reward_fn(
-    scorer: BraintrustScorer,
-    *,
-    messages_to_input: Optional[Callable[[List[Message]], Any]] = None,
-    ground_truth_to_expected: Optional[Callable[[List[Message]], Any]] = None,
-) -> Callable[[List[Message], Optional[List[Message]]], EvaluateResult]:
-    """Wrap a Braintrust scorer as a Reward Kit reward function."""
-
-    @reward_function
-    def reward_fn(
-        messages: List[Message],
-        ground_truth: Optional[List[Message]] = None,
-    ) -> EvaluateResult:
-        input_val = (
-            messages_to_input(messages) if messages_to_input else messages[0].content
-        )
-        output_val = messages[-1].content
-        expected_val = None
-        if ground_truth:
-            expected_val = (
-                ground_truth_to_expected(ground_truth)
-                if ground_truth_to_expected
-                else ground_truth[-1].content
-            )
-        score = scorer(input_val, output_val, expected_val)
-        return EvaluateResult(score=score)
-
-    return reward_fn
-
-
-def reward_fn_to_scorer(
-    reward_fn: Callable[[List[Message], Optional[List[Message]]], EvaluateResult]
-) -> BraintrustScorer:
-    """Create a Braintrust-compatible scorer from a Reward Kit reward function."""
-
-    def scorer(input_val: Any, output: Any, expected: Any) -> float:
-        messages = [
-            Message(role="user", content=str(input_val)),
-            Message(role="assistant", content=str(output)),
-        ]
-        ground_truth = None
-        if expected is not None:
-            ground_truth = [Message(role="assistant", content=str(expected))]
-        result = reward_fn(messages=messages, ground_truth=ground_truth)
-        return result.score
-
-    return scorer
+__all__ = ["scorer_to_reward_fn", "reward_fn_to_scorer"]
