+++ reward_kit/integrations/__init__.py
@@ -1,1 +1,13 @@
+"""Integration helpers for Reward Kit."""
+
+from .openeval import adapt
+from .braintrust import scorer_to_reward_fn, reward_fn_to_scorer
+from .trl import create_trl_adapter
+
+__all__ = [
+    "adapt",
+    "scorer_to_reward_fn",
+    "reward_fn_to_scorer",
+    "create_trl_adapter",
+]
